断句/nike_comment-master/requirements.txt
pytorch==1.11.0
tqdm
boto3
requests
regex
sentencepiece
sacremoses
